fix: coreção dos arquivos pedidos no pr
