+++ apps/mobile/app/recorrencia/index.tsx
@@ -1,62 +1,33 @@
-import { useCallback, useMemo, useState } from 'react'
-import { Pressable, StyleSheet, Text, View } from 'react-native'
-import { router, useFocusEffect } from 'expo-router'
+import { Pressable, View } from 'react-native'
+import { router } from 'expo-router'
 import { Background } from '@/components/ui/Background'
 import { ScrollArea } from '@/components/ui/ScrollArea'
 import { SectionDivider } from '@/components/ui/SectionDivider'
 import { SuccessToast } from '@/components/ui/SuccessToast'
+import { ThemedText } from '@/components/ui/ThemedText'
 import { useThemeColor } from '@/hooks/useThemeColor'
+import { useRecorrenciasScreen } from '@/hooks/useRecorrenciasScreen'
 import { RecurrenceSummaryCard } from '@/components/finance/recurrences/RecurrenceSummaryCard'
 import { RecurrencesList } from '@/components/finance/recurrences/RecurrencesList'
 import { ConfirmacoesPendentes } from '@/components/finance/recurrences/ConfirmacoesPendentes'
-import { mockRecurrences } from '@/components/finance/recurrences/recurrences-data'
-import { consumePendingDeleteId } from '@/components/finance/recurrences/recurrences-store'
 import MaterialIcons from '@expo/vector-icons/MaterialIcons'
+import { indexStyles as styles } from '@/styles/recorrencia.style'
 
 export default function RecorrenciasScreen() {
   const theme = useThemeColor()
-  const [recurrences, setRecurrences] = useState(mockRecurrences)
-  const [toast, setToast] = useState<string | null>(null)
-  const [confirmedIds, setConfirmedIds] = useState<string[]>([])
-  const [skippedIds, setSkippedIds] = useState<string[]>([])
-
-  useFocusEffect(
-    useCallback(() => {
-      const id = consumePendingDeleteId()
-      if (id) {
-        setRecurrences((prev) => prev.filter((r) => r.id !== id))
-        setToast('Recorrência excluída com sucesso.')
-      }
-    }, []),
-  )
-
-  const totalMonthly = useMemo(
-    () =>
-      recurrences
-        .filter((r) => r.isActive && r.frequency === 'MONTHLY')
-        .reduce((sum, r) => (r.type === 'EXPENSE' ? sum + r.amount : sum - r.amount), 0),
-    [recurrences],
-  )
-
-  const activeCount = useMemo(() => recurrences.filter((r) => r.isActive).length, [recurrences])
-
-  const handleToggleActive = (id: string, isActive: boolean) => {
-    setRecurrences((prev) => prev.map((r) => (r.id === id ? { ...r, isActive } : r)))
-  }
-
-  const handleConfirmRecurrence = (id: string) => {
-    setConfirmedIds((prev) => [...prev, id])
-    setToast('Transação registrada com sucesso.')
-  }
-
-  const handleSkipRecurrence = (id: string) => {
-    setSkippedIds((prev) => [...prev, id])
-  }
-
-  const handleUndoRecurrence = (id: string) => {
-    setConfirmedIds((prev) => prev.filter((cid) => cid !== id))
-    setSkippedIds((prev) => prev.filter((sid) => sid !== id))
-  }
+  const {
+    recurrences,
+    toast,
+    setToast,
+    confirmedIds,
+    skippedIds,
+    totalMonthly,
+    activeCount,
+    handleToggleActive,
+    handleConfirmRecurrence,
+    handleSkipRecurrence,
+    handleUndoRecurrence,
+  } = useRecorrenciasScreen()
 
   return (
     <Background>
@@ -69,7 +40,7 @@ export default function RecorrenciasScreen() {
           >
             <MaterialIcons name="arrow-back" size={22} color={theme.foreground} />
           </Pressable>
-          <Text style={[styles.title, { color: theme.foreground }]}>Recorrências</Text>
+          <ThemedText text="Recorrências" style={styles.title} />
           <View style={[styles.iconBtn, { backgroundColor: theme.surfaceMuted }]}>
             <MaterialIcons name="notifications-none" size={22} color={theme.foreground} />
           </View>
@@ -105,51 +76,3 @@ export default function RecorrenciasScreen() {
     </Background>
   )
 }
-
-const styles = StyleSheet.create({
-  scroll: {
-    alignSelf: 'stretch',
-    width: '100%',
-  },
-  content: {
-    paddingHorizontal: 8,
-    paddingTop: 48,
-    paddingBottom: 100,
-    gap: 14,
-    alignItems: 'stretch',
-    width: '100%',
-  },
-  header: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  title: {
-    fontSize: 18,
-    fontWeight: '700',
-    flex: 1,
-    textAlign: 'center',
-  },
-  iconBtn: {
-    width: 38,
-    height: 38,
-    borderRadius: 19,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  fab: {
-    position: 'absolute',
-    bottom: 28,
-    right: 24,
-    width: 56,
-    height: 56,
-    borderRadius: 28,
-    alignItems: 'center',
-    justifyContent: 'center',
-    elevation: 6,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 3 },
-    shadowOpacity: 0.25,
-    shadowRadius: 6,
-  },
-})
