+++ apps/mobile/src/components/finance/recurrences/ConfirmacoesPendentes/style.ts
@@ -1,7 +1,5 @@
 import { StyleSheet } from 'react-native'
 
-const GREEN = '#00E383'
-
 export const styles = StyleSheet.create({
   container: {
     gap: 10,
@@ -27,7 +25,6 @@ export const styles = StyleSheet.create({
     fontSize: 11,
     fontWeight: '700',
     letterSpacing: 0.8,
-    color: GREEN,
   },
   progressLabel: {
     fontSize: 12,
@@ -40,103 +37,7 @@ export const styles = StyleSheet.create({
     overflow: 'hidden',
   },
   progressFill: {
-    backgroundColor: GREEN,
     borderRadius: 2,
-  },
-  card: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 12,
-    borderRadius: 16,
-    paddingHorizontal: 14,
-    paddingVertical: 14,
-    width: '100%',
-    alignSelf: 'stretch',
-  },
-  cardActed: {
-    opacity: 0.55,
-  },
-  iconWrap: {
-    width: 42,
-    height: 42,
-    borderRadius: 21,
-    alignItems: 'center',
-    justifyContent: 'center',
-    flexShrink: 0,
-  },
-  cardBody: {
-    flex: 1,
-    gap: 5,
-    minWidth: 0,
-  },
-  cardTopRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    gap: 8,
-  },
-  cardName: {
-    fontSize: 15,
-    fontWeight: '700',
-    flex: 1,
-  },
-  badge: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 3,
-    flexShrink: 0,
-  },
-  badgeText: {
-    fontSize: 11,
-    fontWeight: '600',
-  },
-  cardMeta: {
-    fontSize: 12,
-    fontWeight: '400',
-  },
-  cardBottomRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    marginTop: 2,
-  },
-  cardAmount: {
-    fontSize: 14,
-    fontWeight: '700',
-  },
-  btnGroup: {
-    flexDirection: 'row',
-    gap: 8,
-  },
-  btn: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 4,
-    paddingHorizontal: 12,
-    paddingVertical: 7,
-    borderRadius: 10,
-  },
-  btnConfirm: {
-    backgroundColor: GREEN,
-  },
-  btnConfirmText: {
-    fontSize: 12,
-    fontWeight: '700',
-    color: '#0F0F13',
-  },
-  btnSkip: {
-    borderWidth: 1,
-  },
-  btnSkipText: {
-    fontSize: 12,
-    fontWeight: '600',
-  },
-  btnUndo: {
-    borderWidth: 1,
-  },
-  btnUndoText: {
-    fontSize: 12,
-    fontWeight: '600',
   },
   allDoneBanner: {
     flexDirection: 'row',
