+++ apps/mobile/src/components/finance/recurrences/DeleteRecurrenceModal/index.tsx
@@ -1,14 +1,11 @@
-import { useState } from 'react'
 import { Modal, Pressable, Text, View } from 'react-native'
 import MaterialIcons from '@expo/vector-icons/MaterialIcons'
+import { ThemedText } from '@/components/ui/ThemedText'
 import { useThemeColor } from '@/hooks/useThemeColor'
-import { formatCurrency } from '@/utils/formatters'
+import { useDeleteRecurrenceModal, type DeleteScope } from '@/hooks/useDeleteRecurrenceModal'
+import { OptionCard } from '../OptionCard'
 import { styles } from './style'
 import type { Recurrence } from '../types'
-
-const RED = '#FF4B4B'
-
-type DeleteScope = 'keep' | 'remove'
 
 type Props = {
   visible: boolean
@@ -17,56 +14,9 @@ type Props = {
   onCancel: () => void
 }
 
-type OptionProps = {
-  selected: boolean
-  onPress: () => void
-  icon: string
-  title: string
-  description: string
-  danger?: boolean
-  theme: ReturnType<typeof useThemeColor>
-}
-
-function OptionCard({ selected, onPress, icon, title, description, danger, theme }: OptionProps) {
-  const accentColor = danger ? RED : theme.primary
-  const borderColor = selected ? accentColor : theme.border
-  const bgColor = selected ? `${accentColor}18` : theme.surfaceMuted
-
-  return (
-    <Pressable onPress={onPress} style={[styles.option, { backgroundColor: bgColor, borderColor }]}>
-      <View style={styles.optionLeft}>
-        <MaterialIcons
-          name={icon as any}
-          size={20}
-          color={selected ? accentColor : theme.mutedForeground}
-        />
-      </View>
-      <View style={styles.optionText}>
-        <Text style={[styles.optionTitle, { color: selected ? accentColor : theme.foreground }]}>
-          {title}
-        </Text>
-        <Text style={[styles.optionDesc, { color: theme.mutedForeground }]}>{description}</Text>
-      </View>
-      <View
-        style={[
-          styles.radio,
-          {
-            borderColor: selected ? accentColor : theme.border,
-            backgroundColor: selected ? accentColor : 'transparent',
-          },
-        ]}
-      >
-        <View style={[styles.radioDot, { display: selected ? 'flex' : 'none' }]} />
-      </View>
-    </Pressable>
-  )
-}
-
 export function DeleteRecurrenceModal({ visible, recurrence, onConfirm, onCancel }: Props) {
   const theme = useThemeColor()
-  const [scope, setScope] = useState<DeleteScope>('keep')
-
-  const amountLabel = recurrence ? formatCurrency(recurrence.amount) : ''
+  const { scope, setScope, amountLabel } = useDeleteRecurrenceModal(recurrence)
 
   return (
     <Modal
@@ -80,17 +30,18 @@ export function DeleteRecurrenceModal({ visible, recurrence, onConfirm, onCancel
           style={[styles.sheet, { backgroundColor: theme.surface, borderColor: theme.border }]}
           onPress={() => {}}
         >
-          <View style={[styles.iconWrap, { backgroundColor: `${RED}22` }]}>
-            <MaterialIcons name="delete-forever" size={28} color={RED} />
+          <View style={[styles.iconWrap, { backgroundColor: `${theme.destructive}22` }]}>
+            <MaterialIcons name="delete-forever" size={28} color={theme.destructive} />
           </View>
 
-          <Text style={[styles.title, { color: theme.foreground }]}>Excluir recorrência</Text>
+          <ThemedText style={styles.title} text="Excluir recorrência" />
           <Text style={[styles.description, { color: theme.mutedForeground }]}>
             Deseja excluir a recorrência{' '}
             <Text style={{ color: theme.foreground, fontWeight: '600' }}>
               "{recurrence?.description ?? ''}"
             </Text>{' '}
-            no valor de <Text style={{ color: RED, fontWeight: '600' }}>{amountLabel}</Text>?
+            no valor de{' '}
+            <Text style={{ color: theme.destructive, fontWeight: '600' }}>{amountLabel}</Text>?
           </Text>
 
           <View style={styles.options}>
@@ -100,7 +51,6 @@ export function DeleteRecurrenceModal({ visible, recurrence, onConfirm, onCancel
               icon="event-available"
               title="Manter transações futuras"
               description="As cobranças agendadas para os próximos meses permanecerão no seu extrato."
-              theme={theme}
             />
             <OptionCard
               selected={scope === 'remove'}
@@ -108,7 +58,6 @@ export function DeleteRecurrenceModal({ visible, recurrence, onConfirm, onCancel
               icon="event-busy"
               title="Remover também as futuras"
               description="Todas as transações vinculadas a esta recorrência serão removidas permanentemente."
-              theme={theme}
               danger
             />
           </View>
@@ -122,19 +71,18 @@ export function DeleteRecurrenceModal({ visible, recurrence, onConfirm, onCancel
                 { borderColor: theme.border, opacity: pressed ? 0.7 : 1 },
               ]}
             >
-              <Text style={[styles.btnText, { color: theme.foreground }]}>Cancelar</Text>
+              <ThemedText style={styles.btnText} text="Cancelar" />
             </Pressable>
 
             <Pressable
               onPress={() => onConfirm(scope)}
               style={({ pressed }) => [
                 styles.btn,
-                styles.btnDelete,
-                { backgroundColor: RED, opacity: pressed ? 0.85 : 1 },
+                { backgroundColor: theme.destructive, opacity: pressed ? 0.85 : 1 },
               ]}
             >
               <MaterialIcons name="delete" size={18} color="#fff" />
-              <Text style={[styles.btnText, { color: '#fff' }]}>Excluir</Text>
+              <ThemedText tone="onPrimary" style={styles.btnText} text="Excluir" />
             </Pressable>
           </View>
         </Pressable>
