+++ apps/mobile/src/components/finance/recurrences/EditScopeModal/index.tsx
@@ -1,10 +1,10 @@
-import { useState } from 'react'
-import { Modal, Pressable, Text, View } from 'react-native'
+import { Modal, Pressable, View } from 'react-native'
 import MaterialIcons from '@expo/vector-icons/MaterialIcons'
+import { ThemedText } from '@/components/ui/ThemedText'
 import { useThemeColor } from '@/hooks/useThemeColor'
+import { useEditScopeModal, type EditScope } from '@/hooks/useEditScopeModal'
+import { OptionCard } from '../OptionCard'
 import { styles } from './style'
-
-export type EditScope = 'upcoming' | 'all'
 
 type Props = {
   visible: boolean
@@ -12,62 +12,11 @@ type Props = {
   onCancel: () => void
 }
 
-type OptionCardProps = {
-  selected: boolean
-  onPress: () => void
-  icon: string
-  title: string
-  description: string
-  theme: ReturnType<typeof useThemeColor>
-}
-
-function OptionCard({ selected, onPress, icon, title, description, theme }: OptionCardProps) {
-  const borderColor = selected ? theme.primary : theme.border
-  const bgColor = selected ? `${theme.primary}18` : theme.surfaceMuted
-
-  return (
-    <Pressable onPress={onPress} style={[styles.option, { backgroundColor: bgColor, borderColor }]}>
-      <View style={styles.optionLeft}>
-        <MaterialIcons
-          name={icon as any}
-          size={20}
-          color={selected ? theme.primary : theme.mutedForeground}
-        />
-      </View>
-      <View style={styles.optionText}>
-        <Text style={[styles.optionTitle, { color: selected ? theme.primary : theme.foreground }]}>
-          {title}
-        </Text>
-        <Text style={[styles.optionDesc, { color: theme.mutedForeground }]}>{description}</Text>
-      </View>
-      <View
-        style={[
-          styles.radio,
-          {
-            borderColor: selected ? theme.primary : theme.border,
-            backgroundColor: selected ? theme.primary : 'transparent',
-          },
-        ]}
-      >
-        <View style={[styles.radioDot, { display: selected ? 'flex' : 'none' }]} />
-      </View>
-    </Pressable>
-  )
-}
+export type { EditScope }
 
 export function EditScopeModal({ visible, onConfirm, onCancel }: Props) {
   const theme = useThemeColor()
-  const [scope, setScope] = useState<EditScope>('upcoming')
-
-  const handleConfirm = () => {
-    onConfirm(scope)
-    setScope('upcoming')
-  }
-
-  const handleCancel = () => {
-    setScope('upcoming')
-    onCancel()
-  }
+  const { scope, setScope, handleConfirm, handleCancel } = useEditScopeModal(onConfirm, onCancel)
 
   return (
     <Modal visible={visible} transparent animationType="fade" onRequestClose={handleCancel}>
@@ -80,12 +29,12 @@ export function EditScopeModal({ visible, onConfirm, onCancel }: Props) {
             <MaterialIcons name="edit" size={28} color={theme.primary} />
           </View>
 
-          <Text style={[styles.title, { color: theme.foreground }]}>
-            Onde aplicar as alterações?
-          </Text>
-          <Text style={[styles.description, { color: theme.mutedForeground }]}>
-            Esta é uma transação recorrente. Escolha como deseja propagar os novos dados.
-          </Text>
+          <ThemedText style={styles.title} text="Onde aplicar as alterações?" />
+          <ThemedText
+            tone="muted"
+            style={styles.description}
+            text="Esta é uma transação recorrente. Escolha como deseja propagar os novos dados."
+          />
 
           <View style={styles.options}>
             <OptionCard
@@ -94,7 +43,6 @@ export function EditScopeModal({ visible, onConfirm, onCancel }: Props) {
               icon="update"
               title="Apenas próximas ocorrências"
               description="Alterações valem de agora em diante. O histórico permanece intacto."
-              theme={theme}
             />
             <OptionCard
               selected={scope === 'all'}
@@ -102,7 +50,6 @@ export function EditScopeModal({ visible, onConfirm, onCancel }: Props) {
               icon="history"
               title="Todas as ocorrências"
               description="Inclui transações passadas vinculadas. Útil para corrigir erros retroativos."
-              theme={theme}
             />
           </View>
 
@@ -115,7 +62,7 @@ export function EditScopeModal({ visible, onConfirm, onCancel }: Props) {
                 { backgroundColor: theme.surfaceMuted, opacity: pressed ? 0.7 : 1 },
               ]}
             >
-              <Text style={[styles.btnText, { color: theme.mutedForeground }]}>Cancelar</Text>
+              <ThemedText tone="muted" style={styles.btnText} text="Cancelar" />
             </Pressable>
 
             <Pressable
@@ -126,9 +73,7 @@ export function EditScopeModal({ visible, onConfirm, onCancel }: Props) {
               ]}
             >
               <MaterialIcons name="check" size={18} color={theme.onPrimary} />
-              <Text style={[styles.btnText, { color: theme.onPrimary }]}>
-                Confirmar e Atualizar
-              </Text>
+              <ThemedText tone="onPrimary" style={styles.btnText} text="Confirmar e Atualizar" />
             </Pressable>
           </View>
         </Pressable>
