+++ apps/mobile/src/components/finance/recurrences/RecurrenceSummaryCard/index.tsx
@@ -1,11 +1,8 @@
 import { Text, View } from 'react-native'
+import { ThemedText } from '@/components/ui/ThemedText'
 import { useThemeColor } from '@/hooks/useThemeColor'
 import { formatCurrency } from '@/utils/formatters'
 import { styles } from './style'
-
-const GREEN = '#00E383'
-const RED = '#FF4B4B'
-const LABEL_GRAY = '#9CA3AF'
 
 type Props = {
   totalMonthly: number
@@ -16,13 +13,13 @@ export function RecurrenceSummaryCard({ totalMonthly, activeCount }: Props) {
   const theme = useThemeColor()
 
   const isExpense = totalMonthly > 0
-  const amountColor = isExpense ? RED : GREEN
+  const amountColor = isExpense ? theme.expense : theme.success
   const sign = isExpense ? '−' : '+'
 
   return (
     <View style={[styles.card, { backgroundColor: theme.surface, borderColor: theme.border }]}>
       <View style={styles.content}>
-        <Text style={[styles.label, { color: LABEL_GRAY }]}>TOTAL MENSAL PREVISTO</Text>
+        <ThemedText tone="muted" style={styles.label} text="TOTAL MENSAL PREVISTO" />
         <Text
           style={[styles.amount, { color: amountColor }]}
           numberOfLines={1}
@@ -30,9 +27,11 @@ export function RecurrenceSummaryCard({ totalMonthly, activeCount }: Props) {
         >
           {sign} {formatCurrency(Math.abs(totalMonthly))}
         </Text>
-        <Text style={[styles.sub, { color: LABEL_GRAY, opacity: activeCount > 0 ? 1 : 0 }]}>
-          {activeCount} recorrências ativas
-        </Text>
+        <ThemedText
+          tone="muted"
+          style={[styles.sub, { opacity: activeCount > 0 ? 1 : 0 }]}
+          text={`${activeCount} recorrências ativas`}
+        />
       </View>
     </View>
   )
